#1383 move prefsFragment to own class
