+++ app/src/main/java/org/kiwix/kiwixmobile/di/components/ApplicationComponent.java
@@ -33,7 +33,7 @@ import org.kiwix.kiwixmobile.downloader.DownloadService;
 import org.kiwix.kiwixmobile.language.LanguageActivity;
 import org.kiwix.kiwixmobile.main.KiwixWebView;
 import org.kiwix.kiwixmobile.search.AutoCompleteAdapter;
-import org.kiwix.kiwixmobile.settings.KiwixSettingsActivity;
+import org.kiwix.kiwixmobile.settings.PrefsFragment;
 import org.kiwix.kiwixmobile.zim_manager.DownloadNotificationClickedReceiver;
 import org.kiwix.kiwixmobile.zim_manager.ZimManageActivity;
 
@@ -66,7 +66,7 @@ public interface ApplicationComponent {
 
   void inject(KiwixWebView kiwixWebView);
 
-  void inject(KiwixSettingsActivity.PrefsFragment prefsFragment);
+  void inject(PrefsFragment prefsFragment);
 
   void inject(AutoCompleteAdapter autoCompleteAdapter);
